2nd part of 3rd chapter's files
create a new folder named *Button*, and move the Button.jsx file previously created to
this folder. Import whatever is needed. 

In the *App.jsx* file, ensure that
import Button from './Button/Button.jsx' is there

In the Button folder, create a new file named
*Button.module.css*

whatever CSS properties are there for button should be transferred to the new
*Button.module.css* file. 

In the *Button.jsx* file, we have to import the module using
*import styles from './Button/Button.module.css'*. Once done, for the className in the
button element, we can put {styles.button}

The nice thing about modules is that it avoids naming conflicts because a unique class
is generated via a hashing algorithm

Modules are convenient if each component has its own unique style.
However, a few downsides with using modules is that it requires additional setup, and
global styles aren't applied easily. We would have to import them from elsewhere